Funcionan de manera basica todos

+++ Practica1/src/practica1/Practica1.java
@@ -9,6 +9,8 @@ import java.util.Scanner;
 public class Practica1 {
 
     public static void main(String[] args) {
+        
+        //Declaraciones
         Scanner scanner = new Scanner(System.in);
         int num;
         LlistaCurs llistaCursos = new LlistaCurs();  // Lista de cursos
@@ -19,9 +21,11 @@ public class Practica1 {
             System.out.println("OPCION 2: Dar de Alta un alumno dada una asignatura a la que se quiere matricular");
             System.out.println("OPCION 3: Dar de baja un curso");
             System.out.println("OPCION 4: Dar de baja una assignatura");
-            System.out.println("OPCION 5: Dado un curso imprimir todas sus asignaturas y los estudiantes que estan matriculados a cada una de ellas");
+            System.out.println("OPCION 5: Ver un curso, sus asignaturas y sus estudiantes");
+            System.out.println("OPCION 6: Dada una asignatura ver el curso y sus estudiantes ");
+            System.out.println("OPCION 7: Dado un estudiante ver que asignaturas tiene y a que curso pertenecen ");
             num = scanner.nextInt();
-            scanner.nextLine();
+            scanner.nextLine(); 
 
             switch (num) {
                 case 1:
@@ -43,7 +47,7 @@ public class Practica1 {
                     System.out.println("Añadir asignaturas");
                     System.out.println("Dime cuántas asignaturas quieres añadir");
                     int numa = scanner.nextInt();
-                    scanner.nextLine();
+                    scanner.nextLine(); 
                     for (int i = 0; i < numa; i++) {
                         Assignatura a;
                         System.out.println("Añade nombre, código y número de créditos");
@@ -57,28 +61,26 @@ public class Practica1 {
                     break;
 
                 case 2:
-                    // Dar de alta un estudiante y matricularlo en una asignatura
-                    Assignatura assignatura;
+                    String codi;
+                    String nom;
+                    String dni;
+                    
+                    
                     Estudiantes estudiante;
-                    System.out.println("Es una asignatura obligatoria(1) o opcional(2)");
-                    int aux = scanner.nextInt();
-                    scanner.nextLine();
-                    if (aux == 1) {
-                        System.out.println("¿Cuál es la asignatura obligatoria a la que te quieres matricular? (Escribe el nombre, el código y los créditos)");
-                        assignatura = new AssignaturaOblig(scanner.nextLine(), scanner.nextLine(), scanner.nextInt());
-                        scanner.nextLine();
-                    } else {
-                        System.out.println("¿Cuál es la asignatura opcional a la que te quieres matricular? (Escribe el nombre, el código y el perfil)");
-                        assignatura = new AssignaturaOpt(scanner.nextLine(), scanner.nextLine(), scanner.nextLine());
-                    }
-                    System.out.println("Escribe el nombre y el DNI del estudiante");
-                    estudiante = new Estudiantes(scanner.nextLine(), scanner.nextLine());
-                    estudiante.matricularAsignatura(assignatura); // Matriculamos el estudiante en la asignatura
-                    listaEstudiantes.addEstudiante(estudiante);  // Añadimos el estudiante a la lista de estudiantes
-                    System.out.println("Estudiante " + estudiante.getNombre() + " con DNI " + estudiante.getDNI() + " matriculado en la asignatura " + assignatura.toString());
+                    
+                    System.out.println("nombre y dni del estudiante a matricular ");
+                    nom=scanner.nextLine();
+                    dni=scanner.nextLine();
+                    estudiante = new Estudiantes(nom, dni);
+                    System.out.println("Que codigo tiene la assignatura a la que queires matricular al alumno?");
+                    codi=scanner.nextLine();
+                    listaEstudiantes.addEstudiante(estudiante);
+                    llistaCursos.BuscarCodiAssignatura(codi, dni);
+                    listaEstudiantes.AfegirAssignaturaAAlumne(dni, codi);
                     break;
 
                 case 3:
+                    
                     // Dar de baja un curso y sus asignaturas
                     System.out.println("Introduce el nombre del curso a dar de baja:");
                     String cursoNombre = scanner.nextLine();
@@ -87,9 +89,16 @@ public class Practica1 {
                         Curso curso = (Curso) llistaCursos.lista().get(i);
                         if (curso.getNombre().equals(cursoNombre)) {
                             // El curso fue encontrado
-                            llistaCursos.lista().remove(i);
+                             for (int j = 0; j < curso.getLlistaAssignatura().longitud(); j++) {
+                                  Assignatura asignatura = curso.getLlistaAssignatura().getObject(j);
+
+                                //borrar codigo de asignatura en Estudiante
+                                String codi_a=asignatura.getIdentificador();
+                                listaEstudiantes.EliminarAsignatura(codi_a);
+                             }
                             encontrado = true;
                             System.out.println("Curso '" + cursoNombre + "' dado de baja correctamente.");
+                            llistaCursos.lista().remove(i);
                             break;
                         }
                     }
@@ -98,24 +107,41 @@ public class Practica1 {
                     }
                     break;
                 case 4:
+                    String error="no entra en el bucle";
+                    //Pedimos datos NOMBRE CURSO
                     System.out.println("Introduce el nombre del curso donde se encuentra la asignatura a dar de baja");
                     cursoNombre = scanner.nextLine();
+                    //Pedimos Dato NOMBRE ASIGNATURA
                     System.out.println("Introduce el nombre de la asignatura a borrar");
-                    String assignaturaNombre = scanner.nextLine();
+                    String assignaturaNombre= scanner.nextLine();
                     encontrado = false;
+                    
                     for (int i = 0; i < llistaCursos.longitud(); i++) {
-                        Curso curso = (Curso) llistaCursos.lista().get(i);
+                        Curso curso = llistaCursos.getCurs(i);
+                        //buscamos el curso en la lista
+                        error="no encuentra un curso igual";
                         if (curso.getNombre().equals(cursoNombre)) {
-                            curso.eliminarAssignatura(assignaturaNombre);
+                            //comprobamos que la asignatura existe en curso
+                            error="no encuentra una assignatura igual";
+                            if(curso.ExisteAssignatura(assignaturaNombre)==true){
+                                encontrado =true;
+                                String codi_a= curso.CodiAssignatura(assignaturaNombre);
+                                listaEstudiantes.EliminarAsignatura(codi_a);
+                                curso.eliminarAssignatura(assignaturaNombre);
+                                error="todo va bien";
+                            }
+                            
                         }
                     }
+                    
+                    //por si no se ha encontrado el curso
                     if (!encontrado) {
-                        System.out.println("Curso no encontrado.");
+                        System.out.println(error);
                     }
-
+                    
                     break;
                 case 5:
-                    String nom_aux,
+                     String nom_aux,
                      codi_aux;
                     System.out.println("Introduce el nombre del curso del cual quieres ver los datos (escribe el nombre y el codigo).");
                     nom_aux = scanner.nextLine();
@@ -137,22 +163,67 @@ public class Practica1 {
 
                                 // Imprimimos los detalles de la asignatura
                                 System.out.println(asignatura.toStringConDetalles());
-
+                                String codi_a=asignatura.getIdentificador();
+                                listaEstudiantes.BuscarAlumnes(codi_a);
+                                
                                 // Imprimimos los estudiantes matriculados en esta asignatura
-                                asignatura.imprimirMatriculados();
+                                
                             }
-                            break;
+                   
+                       
+                            }
+                        }
+                     break;
+                case 6: //DADA UNA ASIGNATURA dar los alumnos y el curso
+                    System.out.println("El nombre de la asignatura es?");
+                    String nombreassignatura= scanner.nextLine();
+                    //buscamos en todos los cursos
+                    for (int i = 0; i < llistaCursos.longitud(); i++) {
+                         Curso curso = (Curso) llistaCursos.lista().get(i);
+                         if(curso.ExisteAssignatura(nombreassignatura)==true){
+                                encontrado =true;
+                                String codi_a= curso.CodiAssignatura(nombreassignatura);
+                                System.out.println(curso.toStringConDetalles());
+                                listaEstudiantes.BuscarAlumnes(codi_a);
+                            }
+                    }
+                    
+                    break;
+                case 7:
+                    String error1="";
+                    System.out.println("Dime el nombre del estudiante");
+                    String nome=scanner.nextLine();
+                    error1="El estudiante no existe";
+                    //buscamos si el estudiante existe
+                    for(int i = 0; i < listaEstudiantes.longitud(); i++){
+                        error1="No hay alumno mismo nombre";
+                        if(listaEstudiantes.getEstudiante(i).getNombre().equals(nome)){
+                            //cogemos todos los codigos del estudiante
+                            for(int j =0;j<listaEstudiantes.getEstudiante(i).getList().size();j++ ){
+                                String codi_a=listaEstudiantes.getEstudiante(i).getAsignaturasMatriculadas(j);
+                                //buscamos en cada curso
+                                
+                                for (int a = 0; a < llistaCursos.longitud(); a++) {
+                                    //buscamos en las asignaturas
+                                    for(int b =0; b < llistaCursos.getCurs(a).getLlistaAssignatura().longitud();b++){
+                                        error1="No hay asignaturas mismo codigoi"; 
+                                        if(llistaCursos.getCurs(a).getLlistaAssignatura().getObject(b).compareCodi(codi_a)==true){
+                                            System.out.println(llistaCursos.getCurs(a).toStringConDetalles());
+                                            System.out.println(llistaCursos.getCurs(a).getLlistaAssignatura().getObject(b).toStringConDetalles());
+                                        }
+                                    }
+                                }
+                            }
                         }
                     }
-
-                    if (!cursoEncontrado) {
-                        System.out.println("No se encontró el curso con el nombre y código proporcionado.");
-                    }
+                    System.out.println(error1);
                     break;
 
+        
             }
-        } while (num != 0);
-
-        scanner.close();  // Cerrar el scanner
-    }
+            }while(num!=0);
+        
+        }
 }
+        
+
